Initial commit of final code for Sudoku Solver using Verilog
The final integrated code developed for the Sudoku Solver game on Cyclone V FPGA using Verilog (bare-metal) using the De1-SoC development kit.

Developed on Intel Quartus with simulations on ModelSim before testing and validation on hardware. Part of the FPGA Design for System-on-Chip module at University of Leeds. Developed as a group project with fully functional code and demonstration completed.